# backend/openapi_final.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Xynova AI",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/roadmap/generate-roadmap":  {
                                                           "post":  {
                                                                        "tags":  [
                                                                                     "roadmap"
                                                                                 ],
                                                                        "summary":  "Generate Roadmap",
                                                                        "operationId":  "generate_roadmap_api_v1_roadmap_generate_roadmap_post",
                                                                        "requestBody":  {
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/TopicCreate"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        },
                                                                                            "required":  true
                                                                                        },
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "Successful Response",
                                                                                                      "content":  {
                                                                                                                      "application/json":  {
                                                                                                                                               "schema":  {
                                                                                                                                                              "$ref":  "#/components/schemas/TaskResponse"
                                                                                                                                                          }
                                                                                                                                           }
                                                                                                                  }
                                                                                                  },
                                                                                          "422":  {
                                                                                                      "description":  "Validation Error",
                                                                                                      "content":  {
                                                                                                                      "application/json":  {
                                                                                                                                               "schema":  {
                                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                          }
                                                                                                                                           }
                                                                                                                  }
                                                                                                  }
                                                                                      },
                                                                        "security":  [
                                                                                         {
                                                                                             "OAuth2PasswordBearer":  [

                                                                                                                      ]
                                                                                         }
                                                                                     ]
                                                                    }
                                                       },
                  "/api/v1/roadmap/":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "roadmap"
                                                                ],
                                                       "summary":  "List Topics",
                                                       "operationId":  "list_topics_api_v1_roadmap__get",
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Successful Response",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "items":  {
                                                                                                                                                           "$ref":  "#/components/schemas/Topic"
                                                                                                                                                       },
                                                                                                                                             "type":  "array",
                                                                                                                                             "title":  "Response List Topics Api V1 Roadmap  Get"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "OAuth2PasswordBearer":  [

                                                                                                     ]
                                                                        }
                                                                    ]
                                                   }
                                       },
                  "/api/v1/roadmap/{topic_id}/graph":  {
                                                           "patch":  {
                                                                         "tags":  [
                                                                                      "roadmap"
                                                                                  ],
                                                                         "summary":  "Update Roadmap Graph",
                                                                         "operationId":  "update_roadmap_graph_api_v1_roadmap__topic_id__graph_patch",
                                                                         "security":  [
                                                                                          {
                                                                                              "OAuth2PasswordBearer":  [

                                                                                                                       ]
                                                                                          }
                                                                                      ],
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "topic_id",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "integer",
                                                                                                               "title":  "Topic Id"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "requestBody":  {
                                                                                             "required":  true,
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/TopicUpdate"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "Successful Response",
                                                                                                       "content":  {
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/Topic"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                   }
                                                                                                   },
                                                                                           "422":  {
                                                                                                       "description":  "Validation Error",
                                                                                                       "content":  {
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     }
                                                       },
                  "/api/v1/roadmap/{topic_id}/confirm":  {
                                                             "post":  {
                                                                          "tags":  [
                                                                                       "roadmap"
                                                                                   ],
                                                                          "summary":  "Confirm Roadmap",
                                                                          "operationId":  "confirm_roadmap_api_v1_roadmap__topic_id__confirm_post",
                                                                          "security":  [
                                                                                           {
                                                                                               "OAuth2PasswordBearer":  [

                                                                                                                        ]
                                                                                           }
                                                                                       ],
                                                                          "parameters":  [
                                                                                             {
                                                                                                 "name":  "topic_id",
                                                                                                 "in":  "path",
                                                                                                 "required":  true,
                                                                                                 "schema":  {
                                                                                                                "type":  "integer",
                                                                                                                "title":  "Topic Id"
                                                                                                            }
                                                                                             }
                                                                                         ],
                                                                          "requestBody":  {
                                                                                              "required":  true,
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/RoadmapConfirmRequest"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          },
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "Successful Response",
                                                                                                        "content":  {
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {

                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                    }
                                                                                                    },
                                                                                            "422":  {
                                                                                                        "description":  "Validation Error",
                                                                                                        "content":  {
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                    }
                                                                                                    }
                                                                                        }
                                                                      }
                                                         },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "TaskResponse":  {
                                                            "properties":  {
                                                                               "task_id":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Task Id"
                                                                                           },
                                                                               "status":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Status"
                                                                                          }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "task_id",
                                                                             "status"
                                                                         ],
                                                            "title":  "TaskResponse"
                                                        },
...
                                       "TopicCreate":  {
                                                           "properties":  {
                                                                              "title":  {
                                                                                            "anyOf":  [
                                                                                                          {
                                                                                                              "type":  "string"
                                                                                                          },
                                                                                                          {
                                                                                                              "type":  "null"
                                                                                                          }
                                                                                                      ],
                                                                                            "title":  "Title"
                                                                                        },
                                                                              "roadmap_graph":  {
                                                                                                    "anyOf":  [
                                                                                                                  {
                                                                                                                      "additionalProperties":  true,
                                                                                                                      "type":  "object"
                                                                                                                  },
                                                                                                                  {
                                                                                                                      "type":  "null"
                                                                                                                  }
                                                                                                              ],
                                                                                                    "title":  "Roadmap Graph"
                                                                                                },
                                                                              "topic":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Topic"
                                                                                        }
                                                                          },
                                                           "type":  "object",
                                                           "required":  [
                                                                            "topic"
                                                                        ],
                                                           "title":  "TopicCreate"
                                                       },
...
                                   },
                       "securitySchemes":  {
                                               "OAuth2PasswordBearer":  {
                                                                            "type":  "oauth2",
                                                                            "flows":  {
                                                                                          "password":  {
                                                                                                           "scopes":  {

                                                                                                                      },
                                                                                                           "tokenUrl":  "api/v1/auth/token"
                                                                                                       }
                                                                                      }
                                                                        }
                                           }
                   }
}